+++ kphp_polyfills.php
@@ -639,9 +639,9 @@ function _php_serialize_helper_run_or_warning(callable $fun) {
 }
 
 function instance_serialize(object $instance): ?string {
-  KPHP\InstanceSerialization\ClassTransformer::$depth = 0;
+  KPHP\MsgPackSerialization\ClassTransformer::$depth = 0;
   return _php_serialize_helper_run_or_warning(static function() use ($instance) {
-    $packer = (new MessagePack\Packer(MessagePack\PackOptions::FORCE_STR))->extendWith(new KPHP\InstanceSerialization\ClassTransformer());
+    $packer = (new MessagePack\Packer(MessagePack\PackOptions::FORCE_STR))->extendWith(new KPHP\MsgPackSerialization\ClassTransformer());
     return $packer->pack($instance);
   });
 }
@@ -650,7 +650,7 @@ function instance_deserialize(string $packed_str, string $type_of_instance): ?ob
   return _php_serialize_helper_run_or_warning(static function() use ($packed_str, $type_of_instance) {
     $unpacked_array = msgpack_deserialize_safe($packed_str);
 
-    $instance_parser = new KPHP\InstanceSerialization\InstanceDeserializer($type_of_instance);
+    $instance_parser = new KPHP\MsgPackSerialization\InstanceDeserializer($type_of_instance);
     return $instance_parser->fromUnpackedArray($unpacked_array);
   });
 }
